01-starter/agent/tool.py
"""
Google Maps MCP Toolset for the Restaurant Agent.

This file creates an MCP (Model Context Protocol) toolset that connects
to the Google Maps API for finding restaurants and places.
"""
import os

from dotenv import load_dotenv
from google.adk.tools.mcp_tool import StreamableHTTPConnectionParams
from google.adk.tools.mcp_tool.mcp_toolset import MCPToolset

MAPS_MCP_URL = "https://mapstools.googleapis.com/mcp"


def get_maps_mcp_toolset():
    """
    Creates and returns the Google Maps MCP toolset.

    The MCP toolset provides access to Google Maps functionality including:
    - Place search
    - Restaurant finding
    - Location details
    """
    load_dotenv()
    maps_api_key = os.getenv('GOOGLE_MAPS_API_KEY', 'no_api_found')

    # TODO: Create the MCPToolset with StreamableHTTPConnectionParams
    # The toolset should:
    # - Use MAPS_MCP_URL as the connection URL
    # - Pass the API key in the headers as "X-Goog-Api-Key"
    #
    # Example structure:
    # tools = MCPToolset(
    #     connection_params=StreamableHTTPConnectionParams(
    #         url=MAPS_MCP_URL,
    #         headers={
    #             "X-Goog-Api-Key": maps_api_key
    #         }
    #     )
    # )
    # REPLACE_MCP_TOOLSET
    tools = None  # REPLACE THIS

    print("MCP Toolset configured for Streamable HTTP connection.")
    return tools
